빌드 중 에러 수정 프롭스를 페이지내에서 해서 문제
